refactor(modal): own the back-close settle in ModalService
Move the "wait for the back-close to settle in the router" logic out of
FilterComponent and into ModalService, where the synthetic history entry
and its history.back() live. componentModal now resolves only once the
overlay has closed and that back-close has settled, so any caller can
navigate on the result without the back reverting it — the filter modal no
longer needs to know about the router or the synthetic entry.
